backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
mcp>=1.13.0
# barrybot: use ChatOpenAI (Mistral base_url) so agentsec OpenAI patcher applies
langchain-openai>=0.2.0
langchain-core>=0.2.0
# agentsec SDK (local aidefense) dependencies
wrapt>=1.14.0
httpx>=0.24.0
pyyaml>=6.0
python-dotenv>=1.0.0
